Umount device instead of label
This fixes GhostBSD/DesktopBSD support.  I suspect this could break USB stick support so this will need testing.  We cannot use a variable to define /media/DESKTOPBSD, and we cannot specify it with an if statement.  Attempting that will allow the media to unmount however pc-sysinstall will not be able to find media for the install.  I don't see another option currently.