Change back to ghostbsd media for now
The /dev/cd0 trick does not work with USB stick.  I will have to find a way to make pc-sysinstall not require the unmounting of /media/GhostBSD.  I am reverting this back to it's original state for now so that I can push the other changes for DesktopBSD before I fall to far behind.